Add a separate WEBHOST flag for browser hosting
Rather than taking a detour via the WEBESSION flag and guessing around
whether the user is allowed to host. Now they need the HOST and WEBHOST
flag, the WEBSESSION flag decides whether they are bound by the
password-only setting or not.

When both guest browser hosting and the ext-auth flag are disabled,
users are now properly disconnected due to lack of permission, rather
than giving them false hope of creating an account and logging in just
to be told they don't have host permission anyway.