Reduce selection resolution to make outlines
When the selection gets too large to feasibly make an outline at full
resolution. This always generates an outline now and never punts to the
mask view, since that was confusing to users.

This also optimizes the common case of a fully opaque (i.e. rectangular)
selection by checking that first, rather than going through the whole
edge detection shebang.