Patch keyboard handling in the browser
It uses a strange off-screen input element that doesn't work properly.
It's now been cobbled together in a way that seems to work on all
platforms I have access to by installing keydown/keyup handlers on the
element. On Android, we still install an input element because otherwise
it inexplicably doesn't trigger key events.