Implement prefer WebSockets server setting
This will make the server indicate that it prefers users connecting via
WebSocket instead of via TCP. Clients will generate invite links with a
"w" query parameter as a result of this. It will also transmit this
preference in the login greeting, in announcements and listings, but at
the time of writing this has no effect.