Implement touch draw pressure, but keep disabled
The implementation works, but depending on the device, the result is
bogus. My Android tablet reports a nonsense pressure that starts at
around 0 and then ramps up to around 0.5 with subsequent events,
regardless of actual pressure applied. My phone just reports full
pressure all the time. So this will need to be a separate setting to
enable to avoid just getting garbage results, but it can't be added yet
because it requires modifying the text, so it will have to be added
after 2.2.2 releases.