Implement inherit and preserve alpha toggles
On layers and tools respectively. By default, they will toggle
themselves automatically based on the blend mode, which is effectively
like it worked previously and is probably what you want most of the
time. That can be toggled off, which also gets rid of the separate
Recolor blend mode, since that's just Normal with alpha preserve.