Fix connection quality being interpreted wrong
The server and client handled it opposite ways round, with the server
interpreting higher numbers as better, but the client sending 0 for bad
quality and 100 for good quality. To fix this, the client now sends
along a new capability indicating that it's doing it correctly and the
server swaps around the numbers for clients that don't send this.