Implement redirect handling in the client
Allowing for transparent or opaque redirects, either immediately during
lookup or, with a bit of a hitch, at session join time. The client will
only accept redirects if the server indicates that it will hand them out
and only actually make a connection if the receiving server indicates it
takes redirects. Additionally, the client passes along an object
received from the source server, which the destination server can use to
only accept signed redirects.

The server part is not yet implemented in a usable way, it's just
confirmed to function.