Implement pixel-perfect option for pixel brushes
Will not put down pixels until the cursor has moved far enough to
determine whether it should be an orthogonal or diagonal one, avoiding
jagged edges. This implementation is probably not ideal, since it
doesn't "preview" where the pixel would be, but it probably still covers
enough use cases to be useful.