Upgrade to Android 35 using Krita Qt5
Targeting Android 35 will become required in November, so this is a
necessary upgrade. It unfortunately breaks the entire UI due to its
edge-to-edge nonsense, but it can yet be turned off (until 36 breaks it
for good.)

This replaces the use of upstream Qt with Krita's version of it, since
they have a bunch of very useful patches and improvements that fix
several issues we have and add some additional nicities. We still have
to apply our own patches on top of that, but most apply cleanly.

This also starts the implementation of a persistent connection service,
but that's quite a bit more involved than just having a service that
pokes the main loop on occasion, so it's disabled for now.

Also adds a bit of code to support building against Qt6, but that's not
fully functional yet (nor is Qt6 functional on Android, we'll need to
patch around on it.)