Add pixel art input setting
For classic pixel brushes. Disables smoothing, stabilization and spline
interpolation. Causes the brush to immediately put down a pixel as soon
as drawing with the tool is started and whenever the cursor enters
another pixel, rather than attempting to compensate for tablet
inaccuracies and waiting until a whole pixel is moved.