Let clients skip catchup on reconnect
By making them remember the canvas history, user list and ACL state at
the point of disconnect and restoring it if possible. Requires collusion
with the server, which will transmit a "history index" to the client
that they can provide when joining a session, containing the session id,
a last reset time and the history position. The client will increment
the history position with every received message that goes into the
session history and replace the entire index whenever the server tells
it to via a server reply containing a "hidx" key.