Show socket connection type in net status
Since it's a piece of information that ought to go somewhere in the UI
and that is a fine place for it. This also removes some text from the
widget to make it thinner, since the socket protocol should be indicator
enough to not need the long words "connecting" and "connected".