Properly ignore invisible MyPaint dabs
Instead of requiring them to be all of infinitesimally small, have a
hardness of zero and an opacity of zero. That was a bug introduced at
some point during the 2.3 cycle. This also now correcty looks at the
protocol values of hardness, opacity and color alpha, rather than at the
raw floating point numbers, which may end up getting rounded to zero.

Since this bug is out in the wild, those invalid brush dabs are also now
ignored in the painting code, to avoid wasting time generating pointless
masks for them.