Wrap settings with auto-generated files
First steps towards getting rid of the crazy settings implementation
(while it's got a nice interface, it has caused multiple crashing bugs,
compatibility problems, slow compilation and the underlying QSettings
has another host of performance and crash issues.)

This replaces the macro tables and template functions with generated
code and macros to do the binding. This interface isn't as pretty, but
significantly faster to compile, speeding up a clean build by about 10%.

The underlying implementation is now swappable, with current
implementations being an in-memory one and one that delegates to the
current settings system that uses QSettings under the hood. Presumably
we'll add an SQLite implementation as well to replace that.