Consistenize touch draw autodetection
It now no longer attempts to interrogate Android devices on whether they
have a stylus because their response is likely to be lies. It's always
set to "guess" by default now.

The "guess" option is now called "automatic" in the settings, with the
currently active option in parentheses afterwards. That makes it clearer
what it's about.

Lastly, the autodetection is now in the global event filter, rather than
inside of the canvas.